DEV-1710: Fix Markdown rendering compatibility in old WebKit
Introduces an adaptive Markdown renderer that dynamically loads a full-featured component (using `react-markdown` and `shiki`) or a lightweight, custom fallback. This is based on a runtime detection of modern RegExp support. Older WebKit engines (pre-Safari 16.4) throw errors when encountering bundles containing features like named capture groups or lookbehind.

This prevents crashes and ensures basic display of AI assistant responses, addressing issues in older Safari browsers.
